State machine improvements (#94)
* A new look at state machine

* Include basic unittests for items controller

* lets not ignore test folders now

* Add extra details to /items. Minor tweaks here and there. Extra logging.

* Raise plex timeout to 60, same as pd. Move extra data to extended dict.

* Add Orionoid Scraper

* Fixed stream data

* Fix frontend CORS

* Apply black formatting to Orionoid

* realdebrid and state fixes

* Add support for multiple scrapers, disabled orionoid for now...

* Fix frontend after state changes

* Add missing files

* Workaround for quick shutdown, fixed ongoing season logic

* Couple minor tweaks. Orionoid still disabled for now.

---------
