+++ backend/program/__init__.py
@@ -1,41 +1,31 @@
 """Program main module"""
-import os
 import threading
 import time
 from typing import Optional
 from pydantic import BaseModel, HttpUrl, Field
-from program.symlink import Symlinker
+from program.media.container import MediaItemContainer
 from utils.logger import logger, get_data_path
 from utils.settings import settings_manager
-from program.media import MediaItemContainer
-from program.libraries.plex import Library as Plex
-from program.debrid.realdebrid import Debrid as RealDebrid
+from program.plex import Plex
+from program.plex import PlexConfig
 from program.content import Content
-from program.scrapers import Scraping
 from utils.utils import Pickly
-
-
-# Pydantic models for configuration
-class PlexConfig(BaseModel):
-    user: str
-    token: str
-    address: HttpUrl
-    watchlist: Optional[HttpUrl] = None
+import concurrent.futures
 
 
 class MdblistConfig(BaseModel):
     lists: list[str] = Field(default_factory=list)
-    api_key: str
+    api_key: Optional[str]
     update_interval: int = 80
 
 
 class OverseerrConfig(BaseModel):
-    url: HttpUrl
-    api_key: str
+    url: Optional[HttpUrl]
+    api_key: Optional[str]
 
 
 class RealDebridConfig(BaseModel):
-    api_key: str
+    api_key: Optional[str]
 
 
 class TorrentioConfig(BaseModel):
@@ -53,11 +43,13 @@ class Settings(BaseModel):
     realdebrid: RealDebridConfig
 
 
-class Program:
+class Program(threading.Thread):
     """Program class"""
 
     def __init__(self):
         logger.info("Iceberg initializing...")
+        super().__init__(name="Iceberg")
+        self.running = False
         self.settings = settings_manager.get_all()
         self.media_items = MediaItemContainer(items=[])
         self.data_path = get_data_path()
@@ -66,13 +58,21 @@ class Program:
         self.threads = [
             Content(self.media_items),  # Content must be first
             Plex(self.media_items),
-            RealDebrid(self.media_items),
-            Symlinker(self.media_items),
-            Scraping(self.media_items),
         ]
         logger.info("Iceberg initialized!")
 
+    def run(self):
+        while self.running:
+            with concurrent.futures.ThreadPoolExecutor(
+                max_workers=10, thread_name_prefix="Worker"
+            ) as executor:
+                for item in self.media_items:
+                    executor.submit(item.perform_action)
+            time.sleep(2)
+
     def start(self):
+        self.running = True
+        super().start()
         for thread in self.threads:
             thread.start()
 
@@ -80,3 +80,5 @@ class Program:
         for thread in self.threads:
             thread.stop()
         self.pickly.stop()
+        self.running = False
+        super().join()
