+++ backend/tests/items_test.py
@@ -1,0 +1,30 @@
+from starlette.testclient import TestClient
+from fastapi import FastAPI
+from program.media.state import MediaItemStates
+import controllers.items as items
+from program.media.container import MediaItemContainer
+from unittest.mock import MagicMock
+
+app = FastAPI()
+app.include_router(items.router)
+app.program = MagicMock()
+app.program.media_items = MediaItemContainer(items=[])
+
+client = TestClient(app)
+
+
+def test_get_states():
+    response = client.get("/items/states")
+    assert response.status_code == 200
+    assert response.json() == {
+        "success": True,
+        "states": [state.name for state in MediaItemStates],
+    }
+
+
+def test_get_items():
+    response = client.get("/items/")
+    assert response.status_code == 200
+    assert isinstance(response.json(), dict)
+    assert response.json()["success"] == True
+    assert isinstance(response.json()["items"], list)
