+++ backend/utils/default_settings.json
@@ -22,6 +22,11 @@
   "torrentio": {
     "filter": "sort=qualitysize%7Cqualityfilter=480p,other,scr,cam,unknown"
   },
+  "orionoid": {
+    "api_key": "",
+    "movie_filter": "&audiolanguages=en,ja,kr,fr,es,ge&audiotype=standard,dubbed&limitretry=25",
+    "tv_filter": "&audiolanguages=en,ja,kr,fr,es,ge&audiotype=standard,dubbed&limitretry=25"        
+  },
   "realdebrid": {
     "api_key": ""
   }
